% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF12.R
\name{as_SF12}
\alias{as_SF12}
\title{Coerce an object to an SF12 vector}
\usage{
as_SF12(x)
}
\arguments{
\item{x}{Object to coerce}
}
\description{
Turns an existing object into an SF12-class vector. Currently only works for
SF12 vectors -- i.e., doesn't do anything.
}
