% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF36.R
\name{as_SF36}
\alias{as_SF36}
\title{Coerce an object to an SF36 vector}
\usage{
as_SF36(x)
}
\arguments{
\item{x}{Object to coerce}
}
\description{
Turns an existing object into an SF36-class vector. Currently only works for
SF36 vectors -- i.e., doesn't do anything.
}
