% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF6D.R
\name{as_SF6D}
\alias{as_SF6D}
\title{Coerce an object to an SF6D vector}
\usage{
as_SF6D(x)
}
\arguments{
\item{x}{Object to coerce}
}
\description{
Turns an existing object into an SF6D-class vector. Currently only works for
SF6D vectors -- i.e., doesn't do anything.
}
