% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{erum}
\alias{erum}
\title{SF-6D modified UK population value set (Episodic Random Utility Model)}
\usage{
erum(x)
}
\arguments{
\item{x}{SF6D vector}
}
\description{
SF-6D modified UK population value set (Episodic Random Utility Model)
}
