% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract dimensions/questions from SF-6D/SF-12 vector}
\usage{
extract(x, dimension)
}
\arguments{
\item{x}{SF6D or SF12 vector}

\item{dimension}{Character string. Names of SF-6D dimension ("PF", "RL",
"SF", "PAIN", "MH", or "VIT") or SF-12 question ("Q1", "Q2", ..., "Q12")
to extract.}
}
\description{
Extract dimensions/questions from SF-6D/SF-12 vector
}
