% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF12.R
\name{is_SF12}
\alias{is_SF12}
\title{Test if the object is an SF12 vector}
\usage{
is_SF12(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
This function returns \code{TRUE} for SF12 vectors (or subclasses thereof)
and \code{FALSE} for all other objects.
}
