% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF36.R
\name{is_SF36}
\alias{is_SF36}
\title{Test if the object is an SF36 vector}
\usage{
is_SF36(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
This function returns \code{TRUE} for SF36 vectors (or subclasses thereof)
and \code{FALSE} for all other objects.
}
