% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF6D.R
\name{is_SF6D}
\alias{is_SF6D}
\title{Test if the object is an SF6D vector}
\usage{
is_SF6D(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
This function returns \code{TRUE} for SF6D vectors (or subclasses thereof)
and \code{FALSE} for all other objects.
}
