% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{Mapping to predict other utility instrument value sets from the SF-6D}
\usage{
mapping(x, instrument, algorithm)
}
\arguments{
\item{x}{\code{SF6D} vector}

\item{instrument}{Which instrument to map to (currently "EQ-5D-5L",
"EQ-5D-3L", "HUI2", "HUI3", "15D", "QWB", and "AQoL-8D" are
implemented).}

\item{algorithm}{Which published mapping algorithm to use. See Details for
the available algorithms for each target instrument.}
}
\description{
Calculate predicted utility values based on other utility instruments from
the SF-6D, based on published mapping algorithms.
}
\details{
The currently implemented mapping algorithms are:

Gamst-Klaussen et al. 2016 ("gk"): non-linear mapping to EQ-5D-5L.

Rowen et al. 2012 ("rowen'): linear mapping via VAS to EQ-5D-3L & HUI2.

Richardson et al. 2015 ("richardson"): symmetrical geometric mean square
regression mapping to EQ-5D-5L, HUI3, 15D, QWB, and AQoL-8D.
}
