% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{portugal}
\alias{portugal}
\title{SF-6D Ferreira et al. Portuguese population value set}
\usage{
portugal(x)
}
\arguments{
\item{x}{SF6D vector}
}
\description{
SF-6D Ferreira et al. Portuguese population value set
}
