% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF12.R
\name{SF12}
\alias{SF12}
\title{\code{SF12} vector}
\usage{
SF12(..., version = 2L)
}
\arguments{
\item{...}{Vectors containing SF-12 question reponses; should be named as
\code{\{Q1, Q2, ..., Q12\}}, \code{\{Q1, Q2a, ..., Q7\}}, or lower case
equivalents.}

\item{version}{Which version (1 or 2) of the SF-12 are the data obtained
from?}
}
\description{
This collates SF-12 question responses into an S3 vector to allow nice print
methods and type checking.
}
