% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf12-scores.R
\name{SF12_scores}
\alias{SF12_scores}
\alias{PCS}
\alias{MCS}
\alias{sf12_PCS}
\alias{sf12_MCS}
\title{SF-12 Component Summary scores}
\usage{
PCS(x)

MCS(x)

sf12_PCS(..., version = 2L)

sf12_MCS(..., version = 2L)

SF12_scores(x)
}
\arguments{
\item{x}{\code{SF12}-class vector.}

\item{...}{Vectors containing SF-12 question reponses. See
\code{\link{SF12}} for details.}

\item{version}{Which version (\code{1} or \code{2}) of the SF-12 are the data obtained
from?}
}
\description{
Calculate PCS and MCS (Physical and Mental Component Summary scores) from
SF-12 vector or SF12 responses.

Calculate SF-12 norm-based scale scores. This function returns a tibble
containing the eight health domain scale scores from the SF-12 (on the
standardised 50/10 scale).
}
