% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF36.R
\name{SF36}
\alias{SF36}
\title{\code{SF36} vector}
\usage{
SF36(...)
}
\arguments{
\item{...}{Vectors containing SF-36 question reponses; should be named as
\code{\{Q1, Q2, ..., Q36\}}, \code{\{Q1, Q2, Q3a, ..., Q11d7\}}, or
lower case equivalents.}
}
\description{
This collates SF-36 question responses into an S3 vector to allow nice print
methods and type checking.
}
