% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf36-scores.R
\name{SF36_scores}
\alias{SF36_scores}
\title{SF-36 Scale Scores}
\usage{
SF36_scores(x)
}
\arguments{
\item{x}{\code{SF36}-class vector.}
}
\description{
Calculate SF-36 health domain scale scores. This function returns a tibble
containing the eight health domain scale scores from the SF-36 (on the
standardised 0--100 scale).
}
