% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF6D.R
\name{SF6D}
\alias{SF6D}
\title{\code{SF6D} vector}
\usage{
SF6D(
  PF = integer(),
  RL = integer(),
  SF = integer(),
  PAIN = integer(),
  MH = integer(),
  VIT = integer(),
  version = "SF-12"
)
}
\arguments{
\item{PF, RL, SF, PAIN, MH, VIT}{Vectors containing SF-6D dimension scores.}

\item{version}{Which questionnaire ("SF-12" or "SF-36") are the data derived
from?}
}
\description{
This collates SF-6D dimension scores into an S3 vector to allow nice print
methods and type checking.
}
