% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf6d_profile.R
\name{sf6d_profile}
\alias{sf6d_profile}
\title{Generate SF-6D profile from SF-12/SF-36 questionnaire}
\usage{
sf6d_profile(..., questionnaire = "SF-12", version = 2L)
}
\arguments{
\item{...}{Vectors containing SF-12 question reponses. See
\code{\link{SF12}} for details.}

\item{questionnaire}{Which questionnaire ('SF-12' or 'SF-36') are the data
obtained from. Only \code{questionnaire = 'SF-12'} is implemented at
this stage.}

\item{version}{Which version (\code{1} or \code{2}) of the SF-12 are the data obtained
from?}
}
\description{
Construct an SF-6D profile from SF-12 (to be implemented: SF-36)
question-level responses. This is a simple wrapper around \code{SF12}
and \code{as_SF6D}.
}
