% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sf6d_utility}
\alias{sf6d_utility}
\title{SF-6D utility values}
\usage{
sf6d_utility(x, values = "uk")
}
\arguments{
\item{x}{SF6D vector}

\item{values}{Which value set to use. Currently only 'uk' (Brazier & Roberts
2004 algorithm) is implemented.}
}
\description{
Calculate SF-6D utility values using specified value set/algorithm.
}
