% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF6Dvalues-package.R
\docType{package}
\name{SF6Dvalues-package}
\alias{SF6Dvalues}
\alias{SF6Dvalues-package}
\title{SF6Dvalues: SF-6D Value Set Calculator}
\description{
Calculators for various SF-6D value sets. The package also
    provides tools for calculation of SF-6D profiles, PCS/MCS scores, and
    health domain scores from SF-12 or SF-36 responses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/uo-cmor/SF6Dvalues}
  \item Report bugs at \url{https://github.com/uo-cmor/SF6Dvalues/issues}
}

}
\author{
\strong{Maintainer}: Ross Wilson \email{ross.wilson@otago.ac.nz}

}
\keyword{internal}
