% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SF6Dvalues.R
\docType{package}
\name{SF6Dvalues}
\alias{SF6Dvalues}
\title{SF6Dvalues: A package to compute SF-12/36 scores and SF-6D utility values}
\description{
This package provides various tools for calculation of SF-6D profiles,
PCS/MCS scores, and health domain scores from SF-12 or SF-36 responses.
}
