% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{spain}
\alias{spain}
\title{SF-6D Abellan Perpinan et al. Spanish population value set}
\usage{
spain(x)
}
\arguments{
\item{x}{SF6D vector}
}
\description{
SF-6D Abellan Perpinan et al. Spanish population value set
}
