% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{utility}
\alias{utility}
\title{SF-6D utility values}
\usage{
utility(x, values = "uk")
}
\arguments{
\item{x}{SF6D vector}

\item{values}{Which value set to use. Currently only 'uk' (Brazier & Roberts
2004 algorithm) is implemented.}
}
\description{
Calculate SF-6D utility values from SF-6D profile (stored in an \code{SF6D}
vector).
}
\seealso{
\code{\link{sf6d_utility}} to calculate SF-6D utility values
directly from SF-12/SF-36 responses; \code{\link{uk}} for the UK
population value set.
}
